makefile done In progress fix issues in functions

+++ libft/ft_strncmp.c
@@ -6,13 +6,16 @@
 /*   By: tsofien- <tsofien-@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/07 11:04:34 by tsofien-          #+#    #+#             */
-/*   Updated: 2023/11/10 11:07:42 by tsofien-         ###   ########.fr       */
+/*   Updated: 2023/11/12 20:47:50 by tsofien-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <stdio.h>
+#include "libft.h"
+
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while ((s1[i] != '\0' || s2[i] != '\0') && i < n && n > 0)
